requirements.txt
Django==2.2
djangorestframework==3.11.0
postgres==3.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
Pillow==7.0.0
django-cors-headers==3.2.1
pandas==1.0.3
xlrd==1.2.0